requirements.txt
# For creating the web API
fastapi==0.111.0

# The ASGI server that runs the FastAPI application
uvicorn[standard]==0.29.0

# The core library for creating and validating FHIR R4 resources
fhir.resources==7.1.0

requests==2.31.0
pandas
thefuzz
python-levenshtein
tqdm
python-dotenv
spacy
